flow_id is 32 bits now, cleanup

+++ inc/librist.h
@@ -482,7 +482,7 @@ RIST_API int rist_server_nack_type_set(struct rist_server *ctx, enum rist_nack_t
  * @note Return immediately
  */
 RIST_API int rist_server_start(struct rist_server *ctx,
-	void (*receive_callback)(void *arg, struct rist_peer *peer, uint64_t flow_id, const void *buffer, size_t len, uint16_t src_port, uint16_t dst_port, uint64_t timestamp_ntp, uint32_t flags),
+	void (*receive_callback)(void *arg, struct rist_peer *peer, uint32_t flow_id, const void *buffer, size_t len, uint16_t src_port, uint16_t dst_port, uint64_t timestamp_ntp, uint32_t flags),
 	void *arg);
 
 /**
